pruebas unitarias

+++ AntaraSoft/Antara.API/Controllers/AgrupacionController.cs
@@ -1,4 +1,6 @@
-﻿using Antara.Model.Contracts.Services;
+﻿using Antara.Model;
+using Antara.Model.Contracts.Services;
+using Antara.Model.Dtos;
 using Antara.Model.Entities;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -22,12 +24,22 @@ namespace Antara.API.Controllers
 
         // url: "localhost:8080/api/agrupacion"
         [HttpPost]
-        public async Task<ActionResult> CreateAgrupacionAsync([FromBody] Agrupacion agrupacion)
+        public async Task<ActionResult<AgrupacionDto>> CreateAgrupacionAsync([FromBody] CreateAgrupacionDto agrupacionDto)
         {
             try
             {
-                var newAgrupacion = await _agrupacionService.CreateAgrupacion(agrupacion);
-                return CreatedAtAction("GetAgrupacion", new { id = newAgrupacion.Id }, newAgrupacion);
+                Agrupacion newAgrupacion = new()
+                {
+                    Id = Guid.NewGuid(),
+                    Name = agrupacionDto.Name,
+                    Description = agrupacionDto.Description,
+                    PublicationDate = DateTime.Parse("1900-01-01"),
+                    IsPublished = false,
+                    Type = agrupacionDto.Type,
+                    User_id = agrupacionDto.User_id
+                };
+                await _agrupacionService.CreateAgrupacion(newAgrupacion);
+                return CreatedAtAction("GetAgrupacion", new { id = newAgrupacion.Id }, newAgrupacion.AsDto());
             }
             catch (Exception err)
             {
@@ -40,12 +52,16 @@ namespace Antara.API.Controllers
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<Audio>> GetAgrupacionAsync(long id)
+        public async Task<ActionResult<AgrupacionDto>> GetAgrupacionAsync(Guid id)
         {
             try
             {
-                var audioList = await _agrupacionService.GetAgrupacion(id);
-                return StatusCode(200, audioList);
+                var agrupacion = await _agrupacionService.GetAgrupacion(id);
+                if (agrupacion == null)
+                {
+                    return NotFound();
+                }
+                return StatusCode(200, agrupacion.AsDto());
             }
             catch (Exception err)
             {
@@ -54,11 +70,11 @@ namespace Antara.API.Controllers
         }
 
         [HttpGet("all/{userId}")]
-        public async Task<ActionResult<List<Agrupacion>>> GetAllAgrupacionAsync(long userId)
+        public async Task<ActionResult<List<AgrupacionDto>>> GetAllAgrupacionAsync(Guid userId)
         {
             try
             {
-                var agrupacionList = await _agrupacionService.GetAllAgrupacion(userId);
+                var agrupacionList = (await _agrupacionService.GetAllAgrupacion(userId)).Select(item => item.AsDto());
                 return StatusCode(200, agrupacionList);
             }
             catch (Exception err)
@@ -68,11 +84,23 @@ namespace Antara.API.Controllers
         }
 
         [HttpPut]
-        public async Task<ActionResult<Audio>> UpdateAgrupacionAsync([FromBody]Agrupacion agrupacion)
+        public async Task<ActionResult> UpdateAgrupacionAsync(Guid id, UpdateAgrupacionDto agrupacionDto)
         {
             try
             {
-                await _agrupacionService.UpdateAgrupacion(agrupacion);
+                Agrupacion existingAgrupacion = await _agrupacionService.GetAgrupacion(id);
+                if (existingAgrupacion == null)
+                {
+                    return NotFound();
+                }
+                Agrupacion updatedAgrupacion = existingAgrupacion with
+                {
+                    Name = agrupacionDto.Name,
+                    Description = agrupacionDto.Description,
+                    IsPublished = agrupacionDto.IsPublished,
+                    Type = agrupacionDto.Type,
+                };
+                await _agrupacionService.UpdateAgrupacion(updatedAgrupacion);
                 return StatusCode(200);
             }
             catch (Exception err)
@@ -82,10 +110,15 @@ namespace Antara.API.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<ActionResult<Audio>> DeleteAgrupacionAsync(long id)
+        public async Task<ActionResult> DeleteAgrupacionAsync(Guid id)
         {
             try
             {
+                var existingAgrupacion = await _agrupacionService.GetAgrupacion(id);
+                if (existingAgrupacion == null)
+                {
+                    return NotFound();
+                }
                 await _agrupacionService.DeleteAgrupacion(id);
                 return StatusCode(200);
             }
@@ -96,11 +129,36 @@ namespace Antara.API.Controllers
         }
 
         [HttpPost("/add")]
-        public async Task<ActionResult> AddAudioToAgrupaacion([FromBody]Agrupacion_Audio agrupacion_audio)
+        public async Task<ActionResult> AddAudioToAgrupacion([FromBody]Agrupacion_Audio agrupacion_audio)
         {
             try
             {
+
+                var existingAgrupacion = await _agrupacionService.GetAgrupacion(agrupacion_audio.Agrupacion_id);
+                if(existingAgrupacion == null)
+                {
+                    return NotFound(agrupacion_audio.Agrupacion_id);
+                }
                 await _agrupacionService.AddAudioToAgrupacion(agrupacion_audio);
+                return StatusCode(200);
+            }
+            catch (Exception err)
+            {
+                return StatusCode(500, err.Message);
+            }
+        }
+
+        [HttpDelete("/remove")]
+        public async Task<ActionResult> RemoveAudioFromAgrupaacion(Guid agrupacionId, Guid audioId)
+        {
+            try
+            {
+                var existingAgrupacion = await _agrupacionService.GetAgrupacion(agrupacionId);
+                if (existingAgrupacion == null)
+                {
+                    return NotFound(agrupacionId);
+                }
+                await _agrupacionService.RemoveAudioFromAgrupacion(agrupacionId, audioId);
                 return StatusCode(200);
             }
             catch (Exception err)
